Added Logger to all the methods and formatted.
