--- Config/config.properties
@@ -1,10 +1,0 @@
-URL=jdbc:mysql://localhost:3306/greeting?useSSL=false&useLegacyDatetimeCode=false&serverTimezone=UTC
-USER=root
-PASSWORD=abcd*1234
-
-PORT=5000
-HOST=localhost
-TIMEOUT=1000
-
-#MODEL VALUES: CLIENT SERVER 
-LOGINLOGOUT=CLIENT
