+++ src/Config/config.properties
@@ -1,0 +1,8 @@
+# Application Port
+PORT=5000
+
+# Application Host
+HOST=localhost
+
+# Connection Timeout
+TIMEOUT=1000
