+++ src/View/Generic/package-info.java
@@ -1,0 +1,5 @@
+package View.Generic;
+
+/**
+ * This package contains all the methods and attributes shared between windows.
+ */
