+++ src/View/LogIn/Login.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<Pane maxHeight="768.0" maxWidth="1365.0" minHeight="600.0" minWidth="800.0" prefHeight="600.0" prefWidth="600.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/8.0.181" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.LogIn.LoginController">
+<Pane maxHeight="768.0" maxWidth="1365.0" minHeight="600.0" minWidth="800.0" prefHeight="600.0" prefWidth="600.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.LogIn.LoginController">
     <children>
         <VBox layoutX="-8.0" prefHeight="612.0" prefWidth="400.0" style="-fx-background-color: #3f61bf;" />
         <Pane layoutX="50.0" layoutY="71.0" prefHeight="487.0" prefWidth="327.0">
@@ -32,7 +32,7 @@
                         <Insets left="40.0" right="10.0" />
                     </padding>
                 </TextField>
-                <Button fx:id="btnLogIn" layoutY="341.0" mnemonicParsing="false" onAction="#handleLogInButtonAction" prefHeight="50.0" prefWidth="306.0" text="Login" textFill="#3f61bf">
+                <Button fx:id="btnLogIn" layoutY="341.0" mnemonicParsing="false" onAction="#handleLogInButtonAction" prefHeight="50.0" prefWidth="319.0" text="Login" textFill="#3f61bf">
                     <font>
                         <Font name="Verdana Pro Black" size="22.0" />
                     </font>
