+++ src/View/SignUp/SignUp.fxml
@@ -13,9 +13,9 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="768.0" maxWidth="1365.0" minHeight="600.0" minWidth="800.0" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.181" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.SignUp.SignUpController">
+<AnchorPane maxHeight="768.0" maxWidth="1365.0" minHeight="600.0" minWidth="800.0" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.SignUp.SignUpController">
     <children>
-        <Pane prefHeight="600.0" prefWidth="200.0" style="-fx-background-color: #3f61bf;">
+        <Pane prefHeight="622.0" prefWidth="200.0" style="-fx-background-color: #3f61bf;">
          <children>
                 <ImageView fitHeight="348.0" fitWidth="323.0" layoutY="112.0" pickOnBounds="true" preserveRatio="true">
                     <image>
@@ -134,10 +134,10 @@
             </TextField>
                 <FontAwesomeIcon fill="#3f61bf" glyphName="LOCK" layoutX="13.0" layoutY="366.0" size="2em" text="" />
                 <FontAwesomeIcon fill="#3f61bf" glyphName="LOCK" layoutX="13.0" layoutY="436.0" size="2em" />
-            <Label fx:id="lblName" layoutX="6.0" layoutY="89.0" prefHeight="17.0" prefWidth="338.0" text="ERROR make sure that your name and surname are well written" textFill="RED" />
+            <Label fx:id="lblName" layoutX="6.0" layoutY="89.0" prefHeight="19.0" prefWidth="364.0" text="ERROR make sure that your name and surname are well written" textFill="RED" />
             <Label fx:id="lblAddress" layoutX="7.0" layoutY="165.0" text="ERROR make sure that your Address is well written" textFill="RED" />
-            <Label fx:id="lblZip" layoutX="9.0" layoutY="229.0" prefHeight="39.0" prefWidth="200.0" text="ERROR make sure that &#10;your zip is well written&#10;" textFill="RED" />
-            <Label fx:id="lblPhone" layoutX="220.0" layoutY="229.0" prefHeight="39.0" prefWidth="152.0" text="ERROR make sure that &#10;your phone is well written" textFill="RED" />
+            <Label fx:id="lblZip" layoutX="9.0" layoutY="229.0" prefHeight="39.0" prefWidth="184.0" text="ERROR make sure that &#10;your zip is well written&#10;" textFill="RED" />
+            <Label fx:id="lblPhone" layoutX="220.0" layoutY="229.0" prefHeight="39.0" prefWidth="184.0" text="ERROR make sure that &#10;your phone is well written" textFill="RED" />
             <Label fx:id="lblEmail" layoutX="7.0" layoutY="313.0" text="ERROR make sure that your email is well written" textFill="RED" />
             <Label fx:id="lblPassword" layoutX="6.0" layoutY="385.0" text="ERROR this password is not safe enough" textFill="RED" />
             <Label fx:id="lblConfirmPassword" layoutX="6.0" layoutY="457.0" text="The password and the confirmation do not match" textFill="RED" />
